Finished weather parsers, split and correct weather services, add 3-hd service in method computeAverageData

+++ src/main/java/com/example/weatherbot/app/dto/weatherapidto/CurrentWeatherInfo.java
@@ -19,7 +19,7 @@ public class CurrentWeatherInfo {
     private Double windDeg;
 
     @JsonProperty("pressure_mb")
-    private Double pressure;
+    private Integer pressure;
 
     @JsonProperty("humidity")
     private Integer humidity;
